Run Docker integration tests with gradle check
build-common.gradle now includes a docker section for starting
docker, running the integration tests in docker and then killing
docker again and cleaning up its files. The integrationTest
is no longer part of gradle check, because it is not expected to
succeed without the help of docker. gradle check now takes about
20 minutes.